requirements.txt
cryptography==2.6.1
Django==2.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==5.1.0
pyasn1==0.1.9
pycparser==2.19
pycryptodome==3.8.1
PyJWT==1.3.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
urllib3==1.13.1
